fix(kernel): remove #![allow(clippy::all)], fix all real clippy warnings
Removes the blanket clippy::all suppression from tylluan-kernel's lib.rs
and main.rs per the external audit's architecture-debt findings. Applied
mechanical auto-fixes (uninlined format args) across tylluan-kernel and
tylluan-link, then manually fixed the remaining ~30 real warnings:
manual clamp -> .clamp(), redundant if/else branches, loop-index-only
for loops -> iter_mut()/enumerate(), a misattributed floating doc comment,
a missing is_empty() companion to len(), a Cow::to_owned() that only
cloned the enum wrapper (-> into_owned()), an unused function parameter,
and a genuine test logic bug (assert!(x || true) in bootstrap.rs, always
passes regardless of x -- replaced with the real intended assertion).
Also removed 4 vacuous is_empty() assertions on compile-time array
literals in dashboard_api_audit.rs and collapsed a duplicated `if` branch
in consensus_stress_test.rs.

Two functions with legitimate many-argument signatures (spawn_heartbeat,
start_http_server_with_download, add_edge_with_validity) got narrow
#[allow(clippy::too_many_arguments)] with a comment explaining why,
rather than forcing a config-struct refactor mid-cleanup.

cargo clippy -p tylluan-kernel -p tylluan-link --all-targets: 0 warnings.
368 kernel + 61 link lib tests pass, plus the touched integration tests
(dashboard_api_audit, http_endpoint_tests, consensus_stress_test).

+++ crates/tylluan-kernel/src/transport/http/api_v1/api_memory.rs
@@ -31,7 +31,7 @@ pub async fn memory_retention_handler(State(state): State<Arc<HttpState>>) -> im
     let silva = state.silva.clone();
     let memory = state.memory.clone();
     let node_count = silva.node_count().await.unwrap_or(0) as i64;
-    let edge_count = silva.edge_count().await.unwrap_or(0) as i64;
+    let edge_count = silva.edge_count().await.unwrap_or(0);
     let memory_stats = memory.stats().await.ok();
     let conn_guard = silva.conn_lock();
     let conn = conn_guard.lock().await;
